Move debounce to service file
 - Add ts annotations
 - Add test for debounce
